requirements.txt
huggingface-hub==0.24.5
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
pillow==10.2.0
pyflakes==2.1.1
PyYAML==6.0.1
regex==2024.5.15
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
setuptools==68.2.2
tb-nightly==2.17.0a20240528
thop==0.1.1.post2209072238
timm==0.9.0
tokenizers==0.19.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm
transformers==4.43.3
yacs==0.1.8
yapf==0.29.0